server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.55.3
